main: Initial implementation of cycle timing / delay
So that the emulation does not go at the speed of light and as actually a bit
usable I've tried to stall (not sleep) the thread emu cycle until a threshold is
met.  The reason we can't sleep / suspend the thread is that SDL is still
capturing keyboard events in the background, and if interrupted would make
inputting key events during the waiting period (which should be most of the
time) impossible.