main: Implement class View, Redo SDL usage
We are now correctly sending a buffer to a texture (I think on the GPU?) instead
of inefficiently coloring NUM_SQUARES SDL_Rects.

Still not 100% certain on the timing method for the emulator.  It appears to be
causing some flickering, and the values that should be for 60Hz are way too slow
so something must be off...