IReader: Initial IReader implementation, change chip8::load sig
In order to test the instructions we need to programatically feed chip8 different
opcodes, and execute them.  Since those fields are (rightfully) private we can't
just instantiate it and go.  Instead we need to 'mock' / 'fake' the file read
call such that in test code we can supply the bytes ourselves, and for real code
it will actually read the file.  So we modify chip8::load's signature to take in
a C++ 'interface' w/ DI as a layer of abstraction to get the control we need.